save output of transformations, handle errors

+++ files/transformations/basic.py
@@ -5,7 +5,7 @@ import zipfile
 import tempfile
 import subprocess
 from ..utils import get_from_s3
-
+from .exceptions import CommandError
 
 def zip_files(*files):
     buffer = io.BytesIO()
@@ -23,15 +23,22 @@ def to_geojson(*files):
     for file in files:
         with open(os.path.join(tmpdir, file.source_filename), "wb") as f:
             f.write(get_from_s3(file).read())
-    subprocess.run(
+    cp = subprocess.run(
         [
             "ogr2ogr",
             "-f",
             "GeoJSON",
             os.path.join(tmpdir, "output.json"),
             os.path.join(tmpdir, files[0].source_filename),
-        ]
+        ],
+        capture_output=True,
+        text=True,
     )
+
+    if cp.returncode != 0:
+        command = " ".join(cp.args)
+        raise CommandError(f"'{command}' returned {cp.returncode}: {cp.stderr}")
+
     with open(os.path.join(tmpdir, "output.json")) as f:
         data = f.read()
     shutil.rmtree(tmpdir)
